+++ src/main/java/com/training/core/config/ConversionConfig.java
@@ -10,8 +10,8 @@ import com.training.core.mapper.model.CargoToCargoResponseConverter;
 import com.training.core.mapper.model.ClientToClientResponseConverter;
 import com.training.core.mapper.model.DeliveryToDeliveryResponseConverter;
 import com.training.core.mapper.model.DeliveryToOrderResponseConverter;
-import com.training.core.mapper.model.PageToDeliveryPageResponse;
-import com.training.core.mapper.model.PageToOrderPageResponseConverter;
+import com.training.core.mapper.model.DeliveryPageToDeliveryPageResponse;
+import com.training.core.mapper.model.OrderPageToOrderPageResponseConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.support.ConversionServiceFactoryBean;
@@ -39,8 +39,11 @@ public class ConversionConfig {
                         toAddressConverter,
                         toCargoConverter,
                         toClientConverter));
-        converters.add(new OrderRequestToDeliveryConverter(
-                toAddressConverter, toCargoConverter, toClientConverter));
+        converters.add(
+                new OrderRequestToDeliveryConverter(
+                        toAddressConverter,
+                        toCargoConverter,
+                        toClientConverter));
 
         AddressToAddressResponceConverter toAddressResponceConverter =
                 new AddressToAddressResponceConverter();
@@ -58,14 +61,17 @@ public class ConversionConfig {
                         toClientResponseConverter);
         converters.add(toDeliveryResponseConverter);
         converters.add(
-                new PageToDeliveryPageResponse(
+                new DeliveryPageToDeliveryPageResponse(
                         toDeliveryResponseConverter));
-        DeliveryToOrderResponseConverter toOrderResponseConverter = new DeliveryToOrderResponseConverter(
-                toAddressResponceConverter,
-                toClientResponseConverter,
-                toCargoResponseConverter);
+        DeliveryToOrderResponseConverter toOrderResponseConverter =
+                new DeliveryToOrderResponseConverter(
+                        toAddressResponceConverter,
+                        toClientResponseConverter,
+                        toCargoResponseConverter);
         converters.add(toOrderResponseConverter);
-        converters.add(new PageToOrderPageResponseConverter(toOrderResponseConverter));
+        converters.add(
+                new OrderPageToOrderPageResponseConverter(
+                        toOrderResponseConverter));
 
         factory.setConverters(converters);
