+++ src/main/java/com/training/core/mapper/model/DeliveryPageToDeliveryPageResponse.java
@@ -1,6 +1,6 @@
 package com.training.core.mapper.model;
 
-import com.training.core.dto.response.OrderPageResponse;
+import com.training.core.dto.response.DeliveryPageResponse;
 import com.training.core.model.Delivery;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.convert.converter.Converter;
@@ -9,19 +9,19 @@ import org.springframework.data.domain.Page;
 import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
-public class PageToOrderPageResponseConverter implements Converter<Page<Delivery>, OrderPageResponse> {
-    private final DeliveryToOrderResponseConverter toOrderResponseConverter;
+public class DeliveryPageToDeliveryPageResponse implements Converter<Page<Delivery>, DeliveryPageResponse> {
+    private final DeliveryToDeliveryResponseConverter toDeliveryResponseConverter;
 
     @Override
-    public OrderPageResponse convert(Page<Delivery> page) {
-        return OrderPageResponse.builder()
+    public DeliveryPageResponse convert(Page<Delivery> page) {
+        return DeliveryPageResponse.builder()
                 .content(page.getContent().stream()
-                        .map(toOrderResponseConverter::convert)
+                        .map(toDeliveryResponseConverter::convert)
                         .collect(Collectors.toList()))
-                .number(page.getNumber())
                 .size(page.getSize())
-                .totalElements(page.getTotalElements())
+                .number(page.getNumber())
                 .totalPages(page.getTotalPages())
+                .totalElements(page.getTotalElements())
                 .build();
     }
